Switch to uv for project management
This change is motivated by the need to update this project's required
Python version. The current version (3.8) is no longer supported by AWS
Elastic Beanstalk and is not supported by newer versions of some project
dependencies.

uv was chosen because it provides good dependency resolution and easy
management of environment variables.

To implement this change, the project's direct dependencies (anything
appearing in an import statement, required by Django, etc.) were pinned
to the versions specified in `requirements.txt`. Then, to help uv
resolve the transitive dependencies, the versions for numpy and
tensorflow-io-gcs-filesystem manually pinned through trial-and-error.

The following (non-transitive) dependencies specified by
`requirements.txt` were determined unnecessary and removed from the
project:

  - astroid
  - bleach
  - chardet
  - decorator
  - dill
  - geocoder
  - gunicorn
  - html5lib
  - isort
  - lazy-object-proxy
  - mccabe
  - platformdirs
  - pylint
  - ratelim
  - tomli
  - tomlkit

The README was updated to reflect the new build process. Some of the
instructions were incomplete (no direction to download the nltk data,
for example) and were updated.